Adapted models to the database data structure.

+++ NomadMapApp/composeApp/src/androidMain/kotlin/org/nomad/mapapp/data/repository/CompanyRepository.kt
@@ -31,6 +31,6 @@ class CompanyRepository(private val apiClient: ApiClient) {
     }
 
     suspend fun rateCompany(companyId: String, score: Float, userId: String): Boolean {
-        return apiClient.rateCompany(companyId, score, userId)
+        return apiClient.rateCompany(companyId.toInt(), score, userId)
     }
 }
